requirements.txt
Django
djangorestframework
flake8
sorl-thumbnail
Pillow